add Select component

+++ src/routes/tasks.ts
@@ -1,25 +1,28 @@
 import { Router } from 'express'
-import { MongoClient, ObjectId } from 'mongodb'
-import type { Task } from '../types'
+import { MongoClient, ObjectId, WithId } from 'mongodb'
+import type { Task, User } from '../types'
 
 export function getTaskRoutes(dbClient: MongoClient): Router {
   const router = Router()
 
   router.post('/', async (req, res) => {
+    const userId = (req.user as WithId<User>)?._id?.toHexString()
     const { title, description }: Omit<Task, 'status' | 'createdAt' | 'updatedAt'> = req.body
 
     const createdAt = new Date().toISOString()
-    const newTask: Task = { title, description, status: 'pending', createdAt, updatedAt: createdAt }
+    const newTask: Task = { title, description, status: 'pending', createdAt, updatedAt: createdAt, userId }
 
     const { insertedId } = await dbClient.db().collection('tasks').insertOne(newTask)
     res.status(201).json({ id: insertedId, ...newTask })
   })
 
   router.get('/', async (req, res) => {
+    const userId = (req.user as WithId<User>)?._id?.toHexString()
+
     const tasks = await dbClient
       .db()
       .collection('tasks')
-      .find()
+      .find({ userId })
       .map(task => ({ ...task, id: task._id.toHexString() }))
       .toArray()
 
@@ -28,11 +31,12 @@ export function getTaskRoutes(dbClient: MongoClient): Router {
 
   router.get('/:id', async (req, res) => {
     const taskId = req.params.id
+    const userId = (req.user as WithId<User>)?._id?.toHexString()
 
     const task = await dbClient
       .db()
       .collection('tasks')
-      .findOne({ _id: new ObjectId(taskId) })
+      .findOne({ _id: new ObjectId(taskId), userId })
 
     if (!task) {
       res.status(404).json({ message: 'Task not found' })
@@ -43,14 +47,16 @@ export function getTaskRoutes(dbClient: MongoClient): Router {
   })
 
   router.patch('/:id', async (req, res) => {
+    const userId = (req.user as WithId<User>)?._id?.toHexString()
     const taskId = req.params.id
+
     const {
       title = '',
       description = '',
       status = 'pending',
-    }: Partial<Omit<Task, 'createdAt' | 'updatedAt'>> = req.body
+    }: Partial<Omit<Task, 'createdAt' | 'updatedAt' | 'userId'>> = req.body
 
-    const updatedTask: Omit<Task, 'createdAt'> = {
+    const updatedTask: Omit<Task, 'createdAt' | 'userId'> = {
       title,
       description,
       status,
@@ -60,7 +66,7 @@ export function getTaskRoutes(dbClient: MongoClient): Router {
     const before = await dbClient
       .db()
       .collection('tasks')
-      .findOneAndUpdate({ _id: new ObjectId(taskId) }, { $set: updatedTask })
+      .findOneAndUpdate({ _id: new ObjectId(taskId), userId }, { $set: updatedTask })
 
     if (!before) {
       res.status(404).json({ message: 'Task not found' })
@@ -71,12 +77,13 @@ export function getTaskRoutes(dbClient: MongoClient): Router {
   })
 
   router.delete('/:id', async (req, res) => {
+    const userId = (req.user as WithId<User>)?._id?.toHexString()
     const taskId = req.params.id
 
     const result = await dbClient
       .db()
       .collection('tasks')
-      .deleteOne({ _id: new ObjectId(taskId) })
+      .deleteOne({ _id: new ObjectId(taskId), userId })
 
     if (result.deletedCount === 0) {
       res.status(404).json({ message: 'Task not found' })
